Module polkadot_collator::service

Structs

RococoParachainRuntimeExecutor

Native executor instance.

SeedlingRuntimeExecutor

Native executor instance.

ShellRuntimeExecutor

Native executor instance.

StatemineRuntimeExecutor

Native Statemine executor instance.

StatemintRuntimeExecutor

Verifier

WaitForAuraConsensus

Special ParachainConsensus implementation that waits for the upgrade from shell to a parachain runtime that implements Aura.

WestmintRuntimeExecutor

Native Westmint executor instance.

Enums

BuildOnAccess

Functions

new_partial

Starts a ServiceBuilder for a full service.

rococo_parachain_build_import_queue

Build the import queue for the rococo parachain runtime.

shell_build_import_queue

Build the import queue for the shell runtime.

start_node_impl

Start a node with the given parachain Configuration and relay chain Configuration.

start_rococo_parachain_node

Start a rococo parachain node.

start_shell_node

Start a polkadot-shell parachain node.

start_shell_node_impl

Start a shell node with the given parachain Configuration and relay chain Configuration.

start_statemint_node

Start a statemint/statemine/westmint parachain node.

statemint_build_import_queue

Build the import queue for the statemint/statemine/westmine runtime.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

Block

Opaque block type.

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

Nonce

Index of a transaction in the chain.